fusion_model

+++ experiments/06_multirate_fusion/train.py
@@ -181,7 +181,7 @@ def train_fold(
     else:
         pos_weight = 1.0
     
-    class_weights = torch.tensor([1.0, pos_weight], device=device)
+    class_weights = torch.tensor([1.0, pos_weight], dtype=torch.float32, device=device)
     criterion = nn.CrossEntropyLoss(weight=class_weights)
     
     # Optimizer
@@ -441,3 +441,4 @@ def main():
 
 if __name__ == "__main__":
     main()
+
